requirements.txt
streamlit
numpy
pillow
cryptography
supabase
matplotlib